Create main.php
More implementation languages coming soon.
I understand this can be done with basic node/express, and I suggest anyone able to do get requests or some form of request handler on user network to handle this tx process.
I am taking any support possible with me on this Algorand project in order to increase productivity. This is a completely non-profit GitHub association.

Please do not hold me responsible I will help and expect others to do so. That is all.